Moved update_active_statuses to its own service class, and moved its tests.  Invoking UpdateUserActiveStatuses at the beginning of monthly emails.
