+++ spec/lib/tasks/user_spec.rb
@@ -4,7 +4,10 @@ require 'rails_helper'
 
 describe 'User rake tasks', type: :task do
   describe ':monthly_stats_email', :inline_jobs do
-    before { allow_any_instance_of(ActorStatsPresenter).to receive(:file_downloads).and_return(3) }
+    before do
+      allow_any_instance_of(ActorStatsPresenter).to receive(:file_downloads).and_return(3)
+      allow(UpdateUserActiveStatuses).to receive(:call).and_return nil
+    end
 
     after { Rake::Task['user:monthly_stats_email'].reenable }
 
@@ -37,72 +40,6 @@ describe 'User rake tasks', type: :task do
         expect {
           Rake::Task['user:monthly_stats_email'].invoke
         }.not_to(change(ActionMailer::Base.deliveries, :count))
-      end
-    end
-  end
-
-  describe ':update_active_statuses', :inline_jobs do
-    let!(:user) { create(:user) }
-
-    after { Rake::Task['user:update_active_statuses'].reenable }
-
-    context 'when user is active faculty' do
-      let(:response) do
-        object_double(PsuIdentity::SearchService::Person.new,
-                      affiliation: ['FACULTY', 'MEMBER'])
-      end
-
-      before do
-        user.active = false
-        user.save!
-        allow_any_instance_of(PsuIdentity::SearchService::Client).to receive(:userid).and_return response
-      end
-
-      it 'updates user active status to true' do
-        Rake::Task['user:update_active_statuses'].invoke
-        user.reload
-        expect(user.active).to eq true
-      end
-    end
-
-    context 'when user is just a MEMBER' do
-      let(:response) do
-        object_double(PsuIdentity::SearchService::Person.new,
-                      affiliation: ['MEMBER'])
-      end
-
-      before do
-        allow_any_instance_of(PsuIdentity::SearchService::Client).to receive(:userid).and_return response
-      end
-
-      it 'updates user active status to false' do
-        Rake::Task['user:update_active_statuses'].invoke
-        user.reload
-        expect(user.active).to eq false
-      end
-    end
-
-    context 'when no user is found' do
-      before do
-        allow_any_instance_of(PsuIdentity::SearchService::Client)
-          .to receive(:userid)
-          .and_raise PsuIdentity::SearchService::NotFound
-      end
-
-      it 'updates user active status to false' do
-        Rake::Task['user:update_active_statuses'].invoke
-        user.reload
-        expect(user.active).to eq false
-      end
-    end
-
-    context 'when timeout occurs' do
-      before do
-        allow_any_instance_of(PsuIdentity::SearchService::Client).to receive(:userid).and_raise Net::ReadTimeout
-      end
-
-      it 'does not change user active status' do
-        expect { Rake::Task['user:update_active_statuses'].invoke }.not_to change(user, :active)
       end
     end
   end
