+++ spec/services/update_user_active_statuses_spec.rb
@@ -1,0 +1,67 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+describe UpdateUserActiveStatuses do
+  let!(:user) { create(:user) }
+
+  context 'when user is active faculty' do
+    let(:response) do
+      object_double(PsuIdentity::SearchService::Person.new,
+                    affiliation: ['FACULTY', 'MEMBER'])
+    end
+
+    before do
+      user.active = false
+      user.save!
+      allow_any_instance_of(PsuIdentity::SearchService::Client).to receive(:userid).and_return response
+    end
+
+    it 'updates user active status to true' do
+      described_class.call
+      user.reload
+      expect(user.active).to eq true
+    end
+  end
+
+  context 'when user is just a MEMBER' do
+    let(:response) do
+      object_double(PsuIdentity::SearchService::Person.new,
+                    affiliation: ['MEMBER'])
+    end
+
+    before do
+      allow_any_instance_of(PsuIdentity::SearchService::Client).to receive(:userid).and_return response
+    end
+
+    it 'updates user active status to false' do
+      described_class.call
+      user.reload
+      expect(user.active).to eq false
+    end
+  end
+
+  context 'when no user is found' do
+    before do
+      allow_any_instance_of(PsuIdentity::SearchService::Client)
+          .to receive(:userid)
+                  .and_raise PsuIdentity::SearchService::NotFound
+    end
+
+    it 'updates user active status to false' do
+      described_class.call
+      user.reload
+      expect(user.active).to eq false
+    end
+  end
+
+  context 'when timeout occurs' do
+    before do
+      allow_any_instance_of(PsuIdentity::SearchService::Client).to receive(:userid).and_raise Net::ReadTimeout
+    end
+
+    it 'does not change user active status' do
+      expect { described_class.call }.not_to change(user, :active)
+    end
+  end
+end
